pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "google-genai"
version = "0.0.1"
description = "GenAI Python SDK"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    { name = "Google LLC", email = "googleapis-packages@google.com" },
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "google-auth>=2.14.1, <3.0.0dev",
    "pillow>=10.0.0, <12.0.0",
    "pydantic>=2.0.0, <3.0.0dev",
    "requests>=2.28.1, <3.0.0dev",
    "websockets>=13.0, <15.0dev",
]

[project.urls]
Homepage = "https://github.com/googlapis/genai-python"

[tool.setuptools]
packages = [
    "google",
    "google.genai",
]
